--- src/components/Header/Header.types.ts
@@ -1,3 +1,0 @@
-export type HeaderProps = {
-  clearTasks: () => void;
-};
