+++ src/components/Tasks/Tasks.tsx
@@ -1,6 +1,10 @@
-import { useEffect, useState } from "react";
-import { placeholders } from "src/content";
-import { TasksProps } from "./Tasks.types";
+import { useEffect, useState, Dispatch } from "react";
+import { placeholders } from "src/content/contents";
+
+export type TasksProps = {
+  tasks: string[];
+  setTasks: Dispatch<string[]>;
+};
 
 export function Tasks({ tasks, setTasks }: TasksProps) {
   const [placeholder, setPlaceholder] = useState<string>("");
