--- src/components/Tasks/index.ts
@@ -1,1 +1,0 @@
-export { Tasks } from "./Tasks";
